feature: session ids for messages
a session document is now being created on backend start up all message documents are updated with a session id that matches the current session.

future updates: add a new chat button and populate a list of chats on the main page. clicking a chat you want will populate with messages that have the same session id that is givin to the chat you select.