update: sideNav chat list updated when a new sessions is created
sessions are not created on startup/refresh, then are only created when "new chat" button is clicked, this now updated a list of chats/sessions and populates on the sideNav.

notes: I think in future sessions should be created when a message/conversation has actually been initiated, or a session will begin but on refresh if there are no messages assosiated with that session it will delete?